+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- added symfony/event-dispatcher ^3.0.0 requirement
+- added optional feature to listen events from RequestSender
 ### Deprecated
 - Method SubscriberInfo::addPropertyChunk is deprecated, use SubscriberInfo::addProperty instead